removed sign up box. changed navbar to add links to homepage and the login page. refactored the code into folders
