+++ src/navigation/NavBar.js
@@ -10,11 +10,17 @@ function NavBar() {
             <nav>
                 <div className="logo"> <img src={logo} alt="UMKC" /> </div>
                 <ul className="nav-links" style={{transform: open ? "translateX(0px)" : "" }}>
-                    <li><a>Home</a></li>
+                    <Link to='/home'>
+                        <li><a>Home</a></li>
+                    </Link>
+                    
                     <li><a>My Account</a></li>
                     <li><a>Submission Portal</a></li>
                     {/*<li><a>About</a></li>*/}
-                    <li><a>Logout</a></li>
+                    <Link to='/'>
+                        <li><a>Logout</a></li>
+                    </Link>
+                    
                     
                 </ul>
                 <i onClick={() => setOpen(!open)} class="fa-solid fa-bars burger" />
